+++ dotnet/Services/CybersourceApi.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Linq;
 using System.Net.Http;
 using System.Security.Cryptography;
 using System.Text;
@@ -65,15 +66,6 @@ namespace Cybersource.Services
                     request.Content = new StringContent(jsonSerializedData, Encoding.UTF8, CybersourceConstants.APPLICATION_JSON);
                 }
 
-                //request.Headers.Add(CybersourceConstants.USE_HTTPS_HEADER_NAME, "true");
-                //string authToken = this._httpContextAccessor.HttpContext.Request.Headers[CybersourceConstants.HEADER_VTEX_CREDENTIAL];
-                //if (authToken != null)
-                //{
-                    //request.Headers.Add(CybersourceConstants.AUTHORIZATION_HEADER_NAME, authToken);
-                    //request.Headers.Add(CybersourceConstants.VTEX_ID_HEADER_NAME, authToken);
-                    //request.Headers.Add(CybersourceConstants.PROXY_AUTHORIZATION_HEADER_NAME, authToken);
-                //}
-
                 string digest = string.Empty;
                 string signatureString = string.Empty;
                 string signature = string.Empty;
@@ -103,14 +95,83 @@ namespace Cybersource.Services
                 };
 
                 Console.WriteLine($"- SendRequest: [{response.StatusCode}] - ");
+
+                StringBuilder sb = new StringBuilder();
+                foreach (var header in request.Headers)
+                {
+                    string headerName = header.Key;
+                    string headerContent = string.Join(",", header.Value.ToArray());
+                    sb.AppendLine($"{headerName} : {headerContent}");
+                    //Console.WriteLine($" |-| {headerName} : {headerContent}");
+                }
+
+                _context.Vtex.Logger.Debug("SendRequest", null, $"{request.RequestUri}\n{sb}\n{jsonSerializedData}\n\n[{response.StatusCode}]\n{responseContent}");
                 //Console.WriteLine($"- SendRequest: [{response.StatusCode}] {responseContent}");
                 //Console.WriteLine($"v-c-merchant-id: {merchantSettings.MerchantId}\nDate: {gmtDateTime}\nHost: {urlBase}\nDigest: {digest}\nSignature: {signatureString}");
 
-                _context.Vtex.Logger.Debug("SendRequest", null, $"{request.RequestUri}\n{jsonSerializedData}\nv-c-merchant-id: {merchantSettings.MerchantId}\nDate: {gmtDateTime}\nHost: {urlBase}\nDigest: {digest}\nSignature: {signatureString}\n[{response.StatusCode}]\n{responseContent}");
+                //_context.Vtex.Logger.Debug("SendRequest", null, $"{request.RequestUri}\n{jsonSerializedData}\nv-c-merchant-id: {merchantSettings.MerchantId}\nDate: {gmtDateTime}\nHost: {urlBase}\nDigest: {digest}\nSignature: {signatureString}\n[{response.StatusCode}]\n{responseContent}");
             }
             catch (Exception ex)
             {
                 _context.Vtex.Logger.Error("SendRequest", null, $"Error ", ex);
+            }
+
+            return sendResponse;
+        }
+
+        public async Task<SendResponse> SendReportRequest(HttpMethod method, string endpoint, string jsonSerializedData)
+        {
+            SendResponse sendResponse = null;
+            MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
+            string urlBase = CybersourceConstants.ProductionApiEndpoint;
+            if (!merchantSettings.IsLive)
+            {
+                urlBase = CybersourceConstants.SandboxApiEndpoint;
+            }
+
+            try
+            {
+                var request = new HttpRequestMessage
+                {
+                    Method = method,
+                    RequestUri = new Uri($"https://{urlBase}{endpoint}")
+                };
+
+                if (!string.IsNullOrEmpty(jsonSerializedData))
+                {
+                    request.Content = new StringContent(jsonSerializedData, Encoding.UTF8, CybersourceConstants.APPLICATION_JSON);
+                }
+
+                string digest = string.Empty;
+                string signatureString = string.Empty;
+                string signature = string.Empty;
+                string gmtDateTime = DateTime.UtcNow.ToString("r");
+
+                request.Headers.Add("v-c-merchant-id", merchantSettings.MerchantId);
+                request.Headers.Add("v-c-date", gmtDateTime);
+                request.Headers.Add("host", urlBase);
+
+                signatureString = await this.GenerateReportSignatureString(merchantSettings, urlBase, gmtDateTime, $"{method.ToString().ToLower()} {endpoint}", digest);
+                request.Headers.Add("signature", signatureString);  // A comma-separated list of parameters that are formatted as name-value pairs
+                request.Headers.Add("accept", "application/hal+json");
+
+                var client = _clientFactory.CreateClient();
+                var response = await client.SendAsync(request);
+                string responseContent = await response.Content.ReadAsStringAsync();
+                sendResponse = new SendResponse
+                {
+                    StatusCode = response.StatusCode.ToString(),
+                    Success = response.IsSuccessStatusCode,
+                    Message = responseContent
+                };
+
+                Console.WriteLine($"- SendReportRequest: [{response.StatusCode}] {responseContent} -");
+
+                _context.Vtex.Logger.Debug("SendReportRequest", null, $"{request.RequestUri}\n{jsonSerializedData}\nv-c-merchant-id: {merchantSettings.MerchantId}\nDate: {gmtDateTime}\nHost: {urlBase}\nDigest: {digest}\nSignature: {signatureString}\n[{response.StatusCode}]\n{responseContent}");
+            }
+            catch (Exception ex)
+            {
+                _context.Vtex.Logger.Error("SendReportRequest", null, $"Error ", ex);
             }
 
             return sendResponse;
@@ -122,11 +183,12 @@ namespace Cybersource.Services
             MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
             //CybersourceToken token = await this.GetOAuthToken(merchantSettings.IsLive);
             string urlBase = CybersourceConstants.ProductionApiEndpoint;
-            string requestUri = $"https://{urlBase}{endpoint}";
             if (!merchantSettings.IsLive)
             {
                 urlBase = CybersourceConstants.SandboxApiEndpoint;
             }
+
+            string requestUri = $"https://{urlBase}{endpoint}";
 
             try
             {
@@ -151,11 +213,20 @@ namespace Cybersource.Services
                 if (!method.Equals(HttpMethod.Get) && !method.Equals(HttpMethod.Delete))
                 {
                     SendResponse proxyTokenSendResponse = await this.SendProxyTokenRequest(jsonSerializedData, proxyTokenUrl);
-                    if(proxyTokenSendResponse.Success)
+                    if (proxyTokenSendResponse.Success)
                     {
                         ProxyTokenResponse proxyToken = JsonConvert.DeserializeObject<ProxyTokenResponse>(proxyTokenSendResponse.Message);
                         digest = proxyToken.Tokens[0].Placeholder;
                     }
+                    else
+                    {
+                        Console.WriteLine("Did not calculate digest!");
+                    }
+
+                    /// TESTING ------------------------------------------
+                    //Console.WriteLine("     !!!!    OVERRIDING CARD DATA FOR TESTING    !!!!    ");
+                    //digest = await this.GenerateDigest(jsonSerializedData);
+                    /// TESTING ------------------------------------------
 
                     request.Headers.Add($"{CybersourceConstants.PROXY_HEADER_PREFIX}Digest", digest); // Do not pass this header field for GET requests. It is a hash of the JSON payload made using a SHA-256 hashing algorithm.
                 }
@@ -184,12 +255,22 @@ namespace Cybersource.Services
                 };
 
                 Console.WriteLine($"- SendRequest: [{response.StatusCode}] - ");
+                StringBuilder sb = new StringBuilder();
+                foreach (var header in request.Headers)
+                {
+                    string headerName = header.Key;
+                    string headerContent = string.Join(",", header.Value.ToArray());
+                    sb.AppendLine($"{headerName} : {headerContent}");
+                    //Console.WriteLine($" |-| {headerName} : {headerContent}");
+                }
 
-                _context.Vtex.Logger.Debug("SendRequest", null, $"{request.RequestUri}\n{jsonSerializedData}\n\n[{response.StatusCode}]\n{responseContent}");
+                _context.Vtex.Logger.Debug("SendRequest", "Proxy", $"{request.RequestUri}\n{sb}\n{jsonSerializedData}\n\n[{response.StatusCode}]\n{responseContent}");
+                //_context.Vtex.Logger.Debug("SendRequest", "Proxy", $"{request.RequestUri}\n{jsonSerializedData}\n\n[{response.StatusCode}]\n{responseContent}");
+                //_context.Vtex.Logger.Debug("SendRequest", "Proxy", $"{request.RequestUri}\n[{response.StatusCode}]\n{responseContent}");
             }
             catch (Exception ex)
             {
-                _context.Vtex.Logger.Error("SendRequest", null, $"Error ", ex);
+                _context.Vtex.Logger.Error("SendRequest", "Proxy", $"Error ", ex);
             }
 
             return sendResponse;
@@ -216,7 +297,7 @@ namespace Cybersource.Services
                         {
                             new RequestToken
                             {
-                                Alias = "digest",
+                                Name = "digest",
                                 Value = new Value
                                 {
                                     Sha256 = new Sha256
@@ -260,13 +341,22 @@ namespace Cybersource.Services
         public async Task<PaymentsResponse> ProcessPayment(Payments payments, string proxyUrl, string proxyTokensUrl)
         {
             PaymentsResponse paymentsResponse = null;
+            /// TESTING ------------------------------------------
             //Console.WriteLine("     !!!!    OVERRIDING CARD DATA FOR TESTING    !!!!    ");
             //payments.paymentInformation.card.number = "4111111111111111";
-            //payments.paymentInformation.card.securityCode = "123";
+            //payments.paymentInformation.card.securityCode = "111";
+            //payments.orderInformation.amountDetails.totalAmount = "401";
+            //payments.orderInformation.billTo.postalCode = "28650";
+            /// TESTING ------------------------------------------
+
             string json = JsonConvert.SerializeObject(payments);
             string endpoint = $"{CybersourceConstants.PAYMENTS}payments";
             SendResponse response = await this.SendProxyRequest(HttpMethod.Post, endpoint, json, proxyUrl, proxyTokensUrl);
+            
+
             //SendResponse response = await this.SendRequest(HttpMethod.Post, endpoint, json);
+            //response = await this.SendRequest(HttpMethod.Post, endpoint, json);
+            
             if (response != null)
             {
                 if (response.Success)
@@ -387,6 +477,138 @@ namespace Cybersource.Services
         }
         #endregion
 
+        #region Reporting
+        /// <summary>
+        /// The Conversion Detail Report contains details of transactions for a merchant.
+        /// To request the report, your client application must send an HTTP GET message to the report server.
+        /// The default format for responses is JSON, but some reports can also return CSV or XML.
+        /// You can set the response to return CSV or XML in the request header by setting the Accept value to either application/xml or text/csv.
+        /// </summary>
+        /// <returns></returns>
+        public async Task<ConversionReportResponse> ConversionDetailReport(DateTime dtStartTime, DateTime dtEndTime)
+        {
+            // https://<url_prefix>/reporting/v3/conversion-details?startTime={startTime}&endTime={endTime}&organizationId={organizationId}
+            // 2016-11-22T12:00:00.000Z
+            ConversionReportResponse retval = null;
+            string startTime = dtStartTime.ToString("yyyy-MM-ddTHH:mm:ssZ");
+            string endTime = dtEndTime.ToString("yyyy-MM-ddTHH:mm:ssZ");
+            //string startTime = dtStartTime.ToString("yyyy-MM-ddT00:00:00.000Z");
+            //string endTime = dtEndTime.ToString("yyyy-MM-ddT23:00:00.000Z");
+            MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
+            string organizationId = merchantSettings.MerchantId;
+            string endpoint = $"{CybersourceConstants.REPORTING}conversion-details?startTime={startTime}&endTime={endTime}&organizationId={organizationId}&";
+            SendResponse response = await this.SendReportRequest(HttpMethod.Get, endpoint, null);
+            if (response != null)
+            {
+                if(response.Success)
+                {
+                    retval = JsonConvert.DeserializeObject<ConversionReportResponse>(response.Message);
+                }
+                else
+                {
+                    ReportErrorResponse reportErrorResponse = JsonConvert.DeserializeObject<ReportErrorResponse>(response.Message);
+                    Console.WriteLine(reportErrorResponse.Message);
+                }
+            }
+
+            return retval;
+        }
+
+        /// <summary>
+        /// Download the Notification of Change report.
+        /// This report shows eCheck-related fields updated as a result of a response to an eCheck settlement transaction.
+        /// </summary>
+        /// <param name="dtStartTime"></param>
+        /// <param name="dtEndTime"></param>
+        /// <returns></returns>
+        public async Task<string> NotificationOfChangesReport(DateTime dtStartTime, DateTime dtEndTime)
+        {
+            string retval = string.Empty;
+            string startTime = dtStartTime.ToString("yyyy-MM-ddTHH:mm:ss.sssZ");
+            string endTime = dtEndTime.ToString("yyyy-MM-ddTHH:mm:ss.sssZ");
+            string endpoint = $"{CybersourceConstants.REPORTING}notification-of-changes&startTime={startTime}&endTime={endTime}";
+            SendResponse response = await this.SendRequest(HttpMethod.Get, endpoint, null);
+            if (response != null)
+            {
+                //= JsonConvert.DeserializeObject<>(response.Message);
+                Console.WriteLine($"[{response.StatusCode}] {response.Message}");
+                retval = $"[{response.StatusCode}] {response.Message}";
+            }
+
+            return retval;
+        }
+
+        /// <summary>
+        /// Download a report using the unique report name and date.
+        /// </summary>
+        /// <param name="dtReportDate"></param>
+        /// <param name="reportName"></param>
+        /// <returns></returns>
+        public async Task<string> DownloadReport(DateTime dtReportDate, string reportName)
+        {
+            string retval = string.Empty;
+            string reportDate = dtReportDate.ToString("yyyy-MM-ddTHH:mm:ss.sssZ");
+            MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
+            string organizationId = merchantSettings.MerchantId;
+            string endpoint = $"{CybersourceConstants.REPORTING}report-downloads?reportDate={reportDate}&organizationId={organizationId}&reportName={reportName}";
+            SendResponse response = await this.SendRequest(HttpMethod.Get, endpoint, null);
+            if (response != null)
+            {
+                //= JsonConvert.DeserializeObject<>(response.Message);
+                Console.WriteLine($"[{response.StatusCode}] {response.Message}");
+                retval = $"[{response.StatusCode}] {response.Message}";
+            }
+
+            return retval;
+        }
+
+        /// <summary>
+        /// Retrieve a list of the available reports to which you are subscribed.
+        /// This will also give you the reportId value, which you can also use to download a report.
+        /// </summary>
+        /// <param name="dtStartTime"></param>
+        /// <param name="dtEndTime"></param>
+        /// <returns></returns>
+        public async Task<string> RetrieveAvailableReports(DateTime dtStartTime, DateTime dtEndTime)
+        {
+            string retval = string.Empty;
+            string startTime = dtStartTime.ToString("yyyy-MM-ddTHH:mm:ss.sssZ");
+            string endTime = dtEndTime.ToString("yyyy-MM-ddTHH:mm:ss.sssZ");
+            //MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
+            //string organizationId = merchantSettings.MerchantId;
+            //string endpoint = $"{CybersourceConstants.REPORTING}reports&startTime={startTime}&endTime={endTime}&organizationId={organizationId}";
+            string endpoint = $"{CybersourceConstants.REPORTING}reports?startTime={startTime}&endTime={endTime}&timeQueryType=executedTime";
+            SendResponse response = await this.SendRequest(HttpMethod.Get, endpoint, null);
+            if (response != null)
+            {
+                //= JsonConvert.DeserializeObject<>(response.Message);
+                Console.WriteLine($"[{response.StatusCode}] {response.Message}");
+                retval = $"[{response.StatusCode}] {response.Message}";
+            }
+
+            return retval;
+        }
+
+        public async Task<string> GetPurchaseAndRefundDetails(DateTime dtStartTime, DateTime dtEndTime)
+        {
+            string retval = string.Empty;
+            string startTime = dtStartTime.ToString("yyyy-MM-ddTHH:mm:ss.sssZ");
+            string endTime = dtEndTime.ToString("yyyy-MM-ddTHH:mm:ss.sssZ");
+            //MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
+            //string organizationId = merchantSettings.MerchantId;
+            string endpoint = $"{CybersourceConstants.REPORTING}purchase-refund-details?startTime={startTime}&endTime={endTime}";
+            SendResponse response = await this.SendRequest(HttpMethod.Get, endpoint, null);
+            if (response != null)
+            {
+                //= JsonConvert.DeserializeObject<>(response.Message);
+                Console.WriteLine($"[{response.StatusCode}] {response.Message}");
+                retval = $"[{response.StatusCode}] {response.Message}";
+            }
+
+            return retval;
+        }
+        #endregion Reporting
+
         #region Authorization Header functions
         private async Task<string> GenerateDigest(string jsonPayload)
         {
@@ -468,8 +690,50 @@ namespace Cybersource.Services
 
             return signatureHeaderValue.ToString();
         }
+
+        private async Task<string> GenerateReportSignatureString(MerchantSettings merchantSettings, string hostName, string gmtDateTime, string requestTarget, string digest)
+        {
+            var signatureString = new StringBuilder();
+            var signatureHeaderValue = new StringBuilder();
+            string headersString = "host v-c-date (request-target) v-c-merchant-id";
+
+            signatureString.Append('\n');
+            signatureString.Append("host");
+            signatureString.Append(": ");
+            signatureString.Append(hostName);
+            signatureString.Append('\n');
+            signatureString.Append("v-c-date");
+            signatureString.Append(": ");
+            signatureString.Append(gmtDateTime);
+            signatureString.Append('\n');
+            signatureString.Append("(request-target)");
+            signatureString.Append(": ");
+            signatureString.Append(requestTarget);
+            signatureString.Append('\n');
+            signatureString.Append("v-c-merchant-id");
+            signatureString.Append(": ");
+            signatureString.Append(merchantSettings.MerchantId);
+            signatureString.Remove(0, 1);
+
+            var signatureByteString = Encoding.UTF8.GetBytes(signatureString.ToString());
+            var decodedKey = Convert.FromBase64String(merchantSettings.SharedSecretKey);
+            var aKeyId = new HMACSHA256(decodedKey);
+            var hashmessage = aKeyId.ComputeHash(signatureByteString);
+            var base64EncodedSignature = Convert.ToBase64String(hashmessage);
+
+            signatureHeaderValue.Append("keyid=\"" + merchantSettings.MerchantKey + "\"");
+            signatureHeaderValue.Append(", algorithm=\"" + CybersourceConstants.SignatureAlgorithm + "\"");
+            signatureHeaderValue.Append(", headers=\"" + headersString + "\"");
+            signatureHeaderValue.Append(", signature=\"" + base64EncodedSignature + "\"");
+
+            //_context.Vtex.Logger.Debug("GenerateSignatureString", null, signatureString.ToString());
+            //Console.WriteLine(signatureHeaderValue.ToString());
+
+            return signatureHeaderValue.ToString();
+        }
         #endregion
 
+        #region OAuth
         public async Task<CybersourceToken> GetOAuthToken(bool isProduction)
         {
             CybersourceToken token = await _cybersourceRepository.LoadToken(isProduction);
@@ -552,5 +816,6 @@ namespace Cybersource.Services
 
             return token;
         }
+        #endregion OAuth
     }
 }