+++ dotnet/Services/ICybersourcePaymentService.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Threading.Tasks;
 using Cybersource.Models;
 
@@ -13,6 +14,12 @@ namespace Cybersource.Services
         Task<SendAntifraudDataResponse> SendAntifraudData(SendAntifraudDataRequest sendAntifraudDataRequest);
         Task<SendAntifraudDataResponse> GetAntifraudStatus(string id);
 
+        Task<ConversionReportResponse> ConversionDetailReport(DateTime dtStartTime, DateTime dtEndTime);
+        Task<ConversionReportResponse> ConversionDetailReport(string atartTime, string endTime);
+        Task<string> RetrieveAvailableReports(DateTime dtStartTime, DateTime dtEndTime);
+        Task<string> RetrieveAvailableReports(string atartTime, string endTime);
+        Task<string> GetPurchaseAndRefundDetails(DateTime dtStartTime, DateTime dtEndTime);
+        Task<string> GetPurchaseAndRefundDetails(string atartTime, string endTime);
         Task<string> GetAuthUrl();
     }
 }