+++ dotnet/Services/VtexApiService.cs
@@ -263,17 +263,42 @@ namespace Cybersource.Services
             }
             else
             {
+                VtexOrderformTaxConfiguration taxConfiguration = null;
                 dynamic orderConfig = JsonConvert.DeserializeObject(jsonSerializedOrderConfig);
-                VtexOrderformTaxConfiguration taxConfiguration = new VtexOrderformTaxConfiguration
+                try
                 {
+                    string existingTaxConfig = JsonConvert.SerializeObject(orderConfig["taxConfiguration"]);
+                    if (!string.IsNullOrEmpty(existingTaxConfig))
+                    {
+                        taxConfiguration = JsonConvert.DeserializeObject<VtexOrderformTaxConfiguration>(existingTaxConfig);
+                    }
+                    else
+                    {
+                        Console.WriteLine($"Empty tax configuration.");
+                    }
+                }
+                catch(Exception ex)
+                {
+                    Console.WriteLine($"Error getting existing config: '{ex.Message}'\n[{orderConfig["taxConfiguration"]}]");
+                }
 
-                };
+                if (taxConfiguration != null && taxConfiguration.Url.Contains("cybersource"))
+                {
+                    taxConfiguration = new VtexOrderformTaxConfiguration
+                    {
 
-                orderConfig["taxConfiguration"] = JToken.FromObject(taxConfiguration);
+                    };
 
-                jsonSerializedOrderConfig = JsonConvert.SerializeObject(orderConfig);
-                bool success = await this._cybersourceRepository.SetOrderConfiguration(jsonSerializedOrderConfig);
-                retval = success.ToString();
+                    orderConfig["taxConfiguration"] = JToken.FromObject(taxConfiguration);
+
+                    jsonSerializedOrderConfig = JsonConvert.SerializeObject(orderConfig);
+                    bool success = await this._cybersourceRepository.SetOrderConfiguration(jsonSerializedOrderConfig);
+                    retval = success.ToString();
+                }
+                else
+                {
+                    retval = $"Not configured for Cybersouce";
+                }
             }
 
             return retval;
@@ -587,6 +612,7 @@ namespace Cybersource.Services
                             break;
                             break;
                         default:
+                            //await this.ProcessConversions();
                             //_context.Vtex.Logger.Info("ProcessNotification", null, $"State {hookNotification.State} not implemeted.");
                             break;
                     }
@@ -721,6 +747,61 @@ namespace Cybersource.Services
             return success;
         }
 
+        public async Task<string> ProcessConversions()
+        {
+            string results = string.Empty;
+            StringBuilder sb = new StringBuilder();
+            DateTime dtStartTime = DateTime.Now.AddDays(-1);
+            DateTime dtEndTime = DateTime.Now;
+            ConversionReportResponse conversionReport = await _cybersourceApi.ConversionDetailReport(dtStartTime, dtEndTime);
+            if (conversionReport != null)
+            {
+                foreach (ConversionDetail conversionDetail in conversionReport.ConversionDetails)
+                {
+                    //sb.AppendLine($"{conversionDetail.MerchantReferenceNumber} {conversionDetail.OriginalDecision} - {conversionDetail.NewDecision} ");
+                    PaymentData paymentData = await _cybersourceRepository.GetPaymentData(conversionDetail.MerchantReferenceNumber);
+                    if (paymentData != null && paymentData.CreatePaymentResponse != null)
+                    {
+                        if (paymentData.CreatePaymentResponse.Status.Equals(CybersourceConstants.VtexAuthStatus.Undefined))
+                        {
+                            bool updateStatus = true;
+                            paymentData.CreatePaymentResponse.Message = conversionDetail.ReviewerComments;
+                            switch (conversionDetail.NewDecision)
+                            {
+                                case CybersourceConstants.CybersourceDecision.Accept:
+                                    paymentData.CreatePaymentResponse.Status = CybersourceConstants.VtexAuthStatus.Approved;
+                                    break;
+                                case CybersourceConstants.CybersourceDecision.Reject:
+                                    paymentData.CreatePaymentResponse.Status = CybersourceConstants.VtexAuthStatus.Denied;
+                                    break;
+                                case CybersourceConstants.CybersourceDecision.Review:
+                                    updateStatus = false;
+                                    break;
+                            }
+
+                            if (updateStatus)
+                            {
+                                if (paymentData.CallbackUrl != null)
+                                {
+                                    SendResponse sendResponse = await this.PostCallbackResponse(paymentData.CallbackUrl, paymentData.CreatePaymentResponse);
+                                    if (sendResponse != null)
+                                    {
+                                        sb.AppendLine($"{conversionDetail.MerchantReferenceNumber} {conversionDetail.OriginalDecision} - {conversionDetail.NewDecision} updated? {sendResponse.Success}");
+                                        Console.WriteLine($"{conversionDetail.MerchantReferenceNumber} {conversionDetail.OriginalDecision} -> {conversionDetail.NewDecision} updated? {sendResponse.Success}");
+                                        await _cybersourceRepository.SavePaymentData(paymentData.PaymentId, paymentData);
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+
+            results = sb.ToString();
+
+            return results;
+        }
+
         public string GetCountryCode(string country)
         {
             return CybersourceConstants.CountryCodesMapping[country];
@@ -744,8 +825,7 @@ namespace Cybersource.Services
                 {
                     new Hook
                     {
-                        //Major = 1,
-                        //Url = new Uri($"https://{this._httpContextAccessor.HttpContext.Request.Headers[TaxjarConstants.HEADER_VTEX_WORKSPACE]}--{this._httpContextAccessor.HttpContext.Request.Headers[TaxjarConstants.VTEX_ACCOUNT_HEADER_NAME]}.myvtex.com/taxjar/oms/invoice")
+
                     }
                 },
                 ItemTaxResponse = new List<ItemTaxResponse>()
@@ -822,6 +902,49 @@ namespace Cybersource.Services
             _context.Vtex.Logger.Info("CybersourceResponseToVtexResponse", null, $"Request: {JsonConvert.SerializeObject(taxResponse)}\nResponse: {JsonConvert.SerializeObject(vtexTaxResponse)}");
 
             return vtexTaxResponse;
+        }
+
+        public async Task<SendResponse> PostCallbackResponse(string callbackUrl, CreatePaymentResponse createPaymentResponse)
+        {
+            SendResponse sendResponse = null;
+
+            if (!string.IsNullOrEmpty(callbackUrl) && createPaymentResponse != null)
+            {
+                try
+                {
+                    var jsonSerializedPaymentResponse = JsonConvert.SerializeObject(createPaymentResponse);
+                    var request = new HttpRequestMessage
+                    {
+                        Method = HttpMethod.Post,
+                        RequestUri = new Uri(callbackUrl),
+                        Content = new StringContent(jsonSerializedPaymentResponse, Encoding.UTF8, CybersourceConstants.APPLICATION_JSON)
+                    };
+
+                    string authToken = this._httpContextAccessor.HttpContext.Request.Headers[CybersourceConstants.HEADER_VTEX_CREDENTIAL];
+                    if (authToken != null)
+                    {
+                        request.Headers.Add(CybersourceConstants.AUTHORIZATION_HEADER_NAME, authToken);
+                    }
+
+                    var client = _clientFactory.CreateClient();
+                    var response = await client.SendAsync(request);
+                    string responseContent = await response.Content.ReadAsStringAsync();
+
+                    sendResponse = new SendResponse
+                    {
+                        Message = responseContent,
+                        StatusCode = response.StatusCode.ToString(),
+                        Success = response.IsSuccessStatusCode
+                    };
+
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine($"PostCallbackResponse {callbackUrl} Error: {ex.Message} InnerException: {ex.InnerException} StackTrace: {ex.StackTrace}");
+                }
+            }
+
+            return sendResponse;
         }
     }
 }