+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Conversion Report
+- Decision Manager
+
 ## [0.1.0] - 2021-09-24
 ### Added